docker: ensure ClickHouse init script is always executed
Recent versions of ClickHouse do not execute the provided entrypoint
script when the database exists. Workaround this by using our own
entrypoint and use it in place of the official one.

See https://github.com/ClickHouse/ClickHouse/pull/50724.

+++ orchestrator/clickhouse/data/docker-entrypoint.sh
@@ -1,10 +1,17 @@
-#!/bin/sh
+#!/bin/bash
 
-set -ex
+set -e
 
-rm -f /tmp/init.sh
-while [ ! -s /tmp/init.sh ]; do
-    sleep 1
-    wget --no-proxy -qO /tmp/init.sh http://akvorado-orchestrator:8080/api/v0/orchestrator/clickhouse/init.sh || continue
-done
-sh /tmp/init.sh
+if [[ $# -lt 1 ]] || [[ "$1" = "--"* ]]; then
+    rm -f /tmp/init.sh
+    while [[ ! -s /tmp/init.sh ]]; do
+        sleep 1
+        echo "Downloading ClickHouse init script..."
+        wget --no-proxy -qO /tmp/init.sh \
+            http://akvorado-orchestrator:8080/api/v0/orchestrator/clickhouse/init.sh || continue
+    done
+    sh /tmp/init.sh
+fi
+
+# Use official entrypoint
+exec /entrypoint.sh "$@"
